+++ Sources/SuperwallKit/Analytics/Superwall Placement/SuperwallEventObjc.swift
@@ -214,6 +214,9 @@ public enum SuperwallEventObjc: Int, CaseIterable {
   /// When the integration attributes are set.
   case integrationAttributes
 
+  /// When a review is requested from the user.
+  case reviewRequested
+
   public init(event: SuperwallEvent) {
     self = event.backingData.objcEvent
   }
@@ -348,6 +351,8 @@ public enum SuperwallEventObjc: Int, CaseIterable {
       return "networkDecoding_fail"
     case .integrationAttributes:
       return "integration_attributes"
+    case .reviewRequested:
+      return "review_requested"
     }
   }
 }