+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.8.2
+
+### Enhancements
+
+- Adds `review_requested` event when a review is requested.
+
+### Fixes
+
+- Fixes rare crash caused by a concurrency issue.
+
 ## 4.8.1
 
 ### Enhancements